Fix signed/unsigned mismatch warnings and remove unused variable
- Change loop variable k to UINT32 in lightmap.cpp for m_NumPoints comparison
- Cast to int in qcsg.cpp for m_NumPoints comparison
- Cast patchnum to int in sparse.cpp comparison
- Change x and numplanes to UINT32 in mathutil.cpp for m_NumPoints comparison
- Change e and e2 to UINT32 in brush.cpp for m_NumPoints comparison
- Add explicit unsigned long casts in transfers.cpp and vismatrixutil.cpp
  for AllocBlock calls to fix size_t to unsigned long warnings
- Remove unused 'delta' variable in lightmap.cpp

+++ src/sdhlt/sdHLRAD/sparse.cpp
@@ -85,7 +85,7 @@ static void		SetVisColumn (int patchnum, bool uncompressedcolumn[MAX_SPARSE_VISM
 			}
 			if (uncompressedcolumn[m]) // visible
 			{
-				if (m < patchnum)
+				if (m < (int)patchnum)
 				{
 					Error ("SetVisColumn: invalid parameter: m < patchnum");
 				}
