+++ src/sdhlt/sdHLRAD/vismatrixutil.cpp
@@ -376,7 +376,7 @@ void            MakeScales(const int threadnum)
         // copy the transfers out
         if (patch->iData)
         {
-			unsigned	data_size = patch->iData * float_size[g_transfer_compress_type] + unused_size;
+			unsigned long	data_size = (unsigned long)(patch->iData * float_size[g_transfer_compress_type] + unused_size);
 
             patch->tData = (transfer_data_t*)AllocBlock(data_size);
             patch->tIndex = CompressTransferIndicies(tIndex_All, patch->iData, &patch->iIndex);
@@ -649,7 +649,7 @@ void            MakeRGBScales(const int threadnum)
         // copy the transfers out
         if (patch->iData)
         {
-			unsigned	data_size = patch->iData * vector_size[g_rgbtransfer_compress_type] + unused_size;
+			unsigned long	data_size = (unsigned long)(patch->iData * vector_size[g_rgbtransfer_compress_type] + unused_size);
 
             patch->tRGBData = (rgb_transfer_data_t*)AllocBlock(data_size);
             patch->tIndex = CompressTransferIndicies(tIndex_All, patch->iData, &patch->iIndex);
